backend/requirements.txt
# Django Core
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Password Hashing
argon2-cffi==23.1.0

# CORS
django-cors-headers==4.3.1

# Environment Variables
python-dotenv==1.0.0

# API Documentation
drf-yasg==1.21.7

# Celery for Background Tasks
celery==5.3.4
redis==5.0.1

# External API Integration
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
pylint==3.0.3
black==23.12.1
